+++ jdbc/src/main/java/com/interface21/jdbc/core/TransactionTemplate.java
@@ -1,0 +1,42 @@
+package com.interface21.jdbc.core;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.function.Function;
+
+import javax.sql.DataSource;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+
+public class TransactionTemplate {
+
+    private final DataSource dataSource;
+
+    public TransactionTemplate(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public <R> R executeTransaction(Function<Connection, R> function) {
+        Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+            R result = function.apply(connection);
+            connection.commit();
+            return result;
+        } catch (SQLException e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    private void rollback(Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException ex) {
+            throw new DataAccessException(ex);
+        }
+    }
+}
